+++ packages/core-components/src/layout/Page/Page.tsx
@@ -25,7 +25,8 @@ const useStyles = makeStyles(() => ({
       "'pageHeader pageHeader pageHeader' 'pageSubheader pageSubheader pageSubheader' 'pageNav pageContent pageSidebar'",
     gridTemplateRows: 'auto auto 1fr',
     gridTemplateColumns: 'auto 1fr auto',
-    minHeight: '100vh',
+    height: '100vh',
+    overflowY: 'auto',
   },
 }));