+++ plugins/catalog-backend/src/next/database/DefaultProcessingDatabase.test.ts
@@ -19,6 +19,7 @@ import { TestDatabaseId, TestDatabases } from '@backstage/backend-test-utils';
 import { Entity, stringifyEntityRef } from '@backstage/catalog-model';
 import { JsonObject } from '@backstage/config';
 import { Knex } from 'knex';
+import { Logger } from 'winston';
 import * as uuid from 'uuid';
 import { DatabaseManager } from './DatabaseManager';
 import { DefaultProcessingDatabase } from './DefaultProcessingDatabase';
@@ -29,12 +30,15 @@ import {
 } from './tables';
 
 describe('Default Processing Database', () => {
-  const logger = getVoidLogger();
+  const defaultLogger = getVoidLogger();
   const databases = TestDatabases.create({
     ids: ['POSTGRES_13', 'POSTGRES_9', 'SQLITE_3'],
   });
 
-  async function createDatabase(databaseId: TestDatabaseId) {
+  async function createDatabase(
+    databaseId: TestDatabaseId,
+    logger: Logger = defaultLogger,
+  ) {
     const knex = await databases.init(databaseId);
     await DatabaseManager.createDatabase(knex);
     return {
@@ -57,6 +61,123 @@ describe('Default Processing Database', () => {
     await db<DbRefreshStateRow>('refresh_state').insert(ref);
   };
 
+  describe('addUprocessedEntities', () => {
+    function mockEntity(name: string, type: string): Entity {
+      return {
+        apiVersion: '1',
+        kind: 'Component',
+        metadata: {
+          name,
+        },
+        spec: {
+          type,
+        },
+      };
+    }
+
+    it.each(databases.eachSupportedId())(
+      'updates refresh state with varying location keys, %p',
+      async databaseId => {
+        const mockWarn = jest.fn();
+        const { db } = await createDatabase(databaseId, ({
+          debug: jest.fn(),
+          warn: mockWarn,
+        } as unknown) as Logger);
+        await db.transaction(async tx => {
+          const knexTx = tx as Knex.Transaction;
+
+          const steps = [
+            {
+              locationKey: undefined,
+              expectedLocationKey: null,
+              type: 'a',
+              expectedType: 'a',
+            },
+            {
+              locationKey: undefined,
+              expectedLocationKey: null,
+              type: 'b',
+              expectedType: 'b',
+            },
+            {
+              locationKey: 'x',
+              expectedLocationKey: 'x',
+              type: 'c',
+              expectedType: 'c',
+            },
+            {
+              locationKey: 'y',
+              expectedLocationKey: 'x',
+              type: 'd',
+              expectedType: 'c',
+              expectConflict: true,
+            },
+            {
+              locationKey: undefined,
+              expectedLocationKey: 'x',
+              type: 'e',
+              expectedType: 'c',
+              expectConflict: true,
+            },
+            {
+              locationKey: 'x',
+              expectedLocationKey: 'x',
+              type: 'f',
+              expectedType: 'f',
+            },
+          ];
+          for (const step of steps) {
+            mockWarn.mockClear();
+
+            await db.addUnprocessedEntities(tx, {
+              sourceKey: 'testing',
+              entities: [
+                {
+                  entity: mockEntity('1', step.type),
+                  locationKey: step.locationKey,
+                },
+              ],
+            });
+
+            if (step.expectConflict) {
+              // eslint-disable-next-line jest/no-conditional-expect
+              expect(mockWarn).toHaveBeenCalledWith(
+                expect.stringMatching(/^Detected conflicting entityRef/),
+              );
+            } else {
+              // eslint-disable-next-line jest/no-conditional-expect
+              expect(mockWarn).not.toHaveBeenCalled();
+            }
+
+            const entities = await knexTx<DbRefreshStateRow>(
+              'refresh_state',
+            ).select();
+            expect(entities).toEqual([
+              expect.objectContaining({
+                entity_ref: 'component:default/1',
+                location_key: step.expectedLocationKey,
+              }),
+            ]);
+            const entity = JSON.parse(entities[0].unprocessed_entity) as Entity;
+            expect(entity.spec?.type).toEqual(step.expectedType);
+
+            await expect(
+              knexTx<DbRefreshStateReferencesRow>(
+                'refresh_state_references',
+              ).select(),
+            ).resolves.toEqual([
+              expect.objectContaining({
+                source_key: 'testing',
+                target_entity_ref: 'component:default/1',
+              }),
+            ]);
+          }
+        });
+      },
+      60_000,
+    );
+  });
+
   describe('updateProcessedEntity', () => {
     let id: string;
     let processedEntity: Entity;
@@ -77,7 +198,7 @@ describe('Default Processing Database', () => {
     });
 
     it.each(databases.eachSupportedId())(
-      'fails when there is no processing state for the entity, %p',
+      'fails when an entity is processed with a different locationKey, %p',
       async databaseId => {
         const { db } = await createDatabase(databaseId);
         await db.transaction(async tx => {
@@ -89,8 +210,51 @@ describe('Default Processing Database', () => {
               relations: [],
               deferredEntities: [],
             }),
-          ).rejects.toThrow(`Processing state not found for ${id}`);
+          ).rejects.toThrow(
+            `Conflicting write of processing result for ${id} with location key 'undefined'`,
+          );
         });
+      },
+      60_000,
+    );
+
+    it.each(databases.eachSupportedId())(
+      'fails when the locationKey is different, %p',
+      async databaseId => {
+        const options = {
+          id,
+          processedEntity,
+          state: new Map<string, JsonObject>(),
+          relations: [],
+          deferredEntities: [],
+          locationKey: 'key',
+          errors: "['something broke']",
+        };
+        const { knex, db } = await createDatabase(databaseId);
+        await insertRefreshStateRow(knex, {
+          entity_id: id,
+          entity_ref: 'location:default/fakelocation',
+          unprocessed_entity: '{}',
+          processed_entity: '{}',
+          location_key: 'key',
+          errors: '[]',
+          next_update_at: '2021-04-01 13:37:00',
+          last_discovery_at: '2021-04-01 13:37:00',
+        });
+        await db.transaction(tx => db.updateProcessedEntity(tx, options));
+
+        const entities = await knex<DbRefreshStateRow>(
+          'refresh_state',
+        ).select();
+        expect(entities.length).toBe(1);
+
+        await db.transaction(tx =>
+          expect(
+            db.updateProcessedEntity(tx, { ...options, locationKey: 'fail' }),
+          ).rejects.toThrow(
+            `Conflicting write of processing result for ${id} with location key 'fail'`,
+          ),
+        );
       },
       60_000,
     );
@@ -119,6 +283,7 @@ describe('Default Processing Database', () => {
             state,
             relations: [],
             deferredEntities: [],
+            locationKey: 'key',
             errors: "['something broke']",
           }),
         );
@@ -132,6 +297,7 @@ describe('Default Processing Database', () => {
         );
         expect(entities[0].cache).toEqual(JSON.stringify(state));
         expect(entities[0].errors).toEqual("['something broke']");
+        expect(entities[0].location_key).toEqual('key');
       },
       60_000,
     );
@@ -206,11 +372,14 @@ describe('Default Processing Database', () => {
 
         const deferredEntities = [
           {
-            apiVersion: '1',
-            kind: 'Location',
-            metadata: {
-              name: 'next',
+            entity: {
+              apiVersion: '1',
+              kind: 'Location',
+              metadata: {
+                name: 'next',
+              },
             },
+            locationKey: 'mock',
           },
         ];
 
@@ -227,7 +396,7 @@ describe('Default Processing Database', () => {
         const refreshStateEntries = await knex<DbRefreshStateRow>(
           'refresh_state',
         )
-          .where({ entity_ref: stringifyEntityRef(deferredEntities[0]) })
+          .where({ entity_ref: stringifyEntityRef(deferredEntities[0].entity) })
           .select();
 
         expect(refreshStateEntries).toHaveLength(1);
@@ -297,12 +466,15 @@ describe('Default Processing Database', () => {
             sourceKey: 'config',
             items: [
               {
-                apiVersion: '1.0.0',
-                metadata: {
-                  name: 'new-root',
-                },
-                kind: 'Location',
-              } as Entity,
+                entity: {
+                  apiVersion: '1.0.0',
+                  metadata: {
+                    name: 'new-root',
+                  },
+                  kind: 'Location',
+                } as Entity,
+                locationKey: 'file:///tmp/foobar',
+              },
             ],
           }),
         );
@@ -411,12 +583,15 @@ describe('Default Processing Database', () => {
             sourceKey: 'config',
             items: [
               {
-                apiVersion: '1.0.0',
-                metadata: {
-                  name: 'new-root',
-                },
-                kind: 'Location',
-              } as Entity,
+                entity: {
+                  apiVersion: '1.0.0',
+                  metadata: {
+                    name: 'new-root',
+                  },
+                  kind: 'Location',
+                } as Entity,
+                locationKey: 'file:/tmp/foobar',
+              },
             ],
           });
         });
@@ -510,12 +685,15 @@ describe('Default Processing Database', () => {
             removed: [],
             added: [
               {
-                apiVersion: '1.0.0',
-                metadata: {
-                  name: 'new-root',
-                },
-                kind: 'Location',
-              } as Entity,
+                entity: {
+                  apiVersion: '1.0.0',
+                  metadata: {
+                    name: 'new-root',
+                  },
+                  kind: 'Location',
+                } as Entity,
+                locationKey: 'file:///tmp/foobar',
+              },
             ],
           });
         });
@@ -614,12 +792,15 @@ describe('Default Processing Database', () => {
             added: [],
             removed: [
               {
-                apiVersion: '1.0.0',
-                metadata: {
-                  name: 'new-root',
-                },
-                kind: 'Location',
-              } as Entity,
+                entity: {
+                  apiVersion: '1.0.0',
+                  metadata: {
+                    name: 'new-root',
+                  },
+                  kind: 'Location',
+                } as Entity,
+                locationKey: 'file:/tmp/foobar',
+              },
             ],
           });
         });
@@ -645,6 +826,73 @@ describe('Default Processing Database', () => {
               t.target_entity_ref === 'location:default/new-root',
           ),
         ).toBeFalsy();
+      },
+      60_000,
+    );
+
+    it.each(databases.eachSupportedId())(
+      'should update the location key during full replace, %p',
+      async databaseId => {
+        const { knex, db } = await createDatabase(databaseId);
+        await createLocations(knex, ['location:default/removed']);
+        await insertRefreshStateRow(knex, {
+          entity_id: uuid.v4(),
+          entity_ref: 'location:default/replaced',
+          unprocessed_entity: '{}',
+          processed_entity: '{}',
+          errors: '[]',
+          next_update_at: '2021-04-01 13:37:00',
+          last_discovery_at: '2021-04-01 13:37:00',
+          location_key: 'file:///tmp/old',
+        });
+
+        await insertRefRow(knex, {
+          source_key: 'lols',
+          target_entity_ref: 'location:default/removed',
+        });
+        await insertRefRow(knex, {
+          source_key: 'lols',
+          target_entity_ref: 'location:default/replaced',
+        });
+
+        await db.transaction(async tx => {
+          await db.replaceUnprocessedEntities(tx, {
+            type: 'full',
+            sourceKey: 'lols',
+            items: [
+              {
+                entity: {
+                  apiVersion: '1.0.0',
+                  metadata: {
+                    name: 'replaced',
+                  },
+                  kind: 'Location',
+                } as Entity,
+                locationKey: 'file:///tmp/foobar',
+              },
+            ],
+          });
+        });
+
+        const currentRefreshState = await knex<DbRefreshStateRow>(
+          'refresh_state',
+        ).select();
+        expect(currentRefreshState).toEqual([
+          expect.objectContaining({
+            entity_ref: 'location:default/replaced',
+            location_key: 'file:///tmp/foobar',
+          }),
+        ]);
+
+        const currentRefRowState = await knex<DbRefreshStateReferencesRow>(
+          'refresh_state_references',
+        ).select();
+        expect(currentRefRowState).toEqual([
+          expect.objectContaining({
+            source_key: 'lols',
+            target_entity_ref: 'location:default/replaced',
+          }),
+        ]);
       },
       60_000,
     );