+++ plugins/scaffolder-backend/src/scaffolder/actions/builtin/publish/bitbucket.ts
@@ -20,9 +20,10 @@ import {
   ScmIntegrationRegistry,
 } from '@backstage/integration';
 import fetch from 'cross-fetch';
-import { initRepoAndPush } from '../../../stages/publish/helpers';
+import { initRepoAndPush } from '../helpers';
 import { createTemplateAction } from '../../createTemplateAction';
 import { getRepoSourceDirectory, parseRepoUrl } from './util';
+import { Config } from '@backstage/config';
 
 const createBitbucketCloudRepository = async (opts: {
   owner: string;
@@ -184,12 +185,14 @@ const performEnableLFS = async (opts: {
 
 export function createPublishBitbucketAction(options: {
   integrations: ScmIntegrationRegistry;
+  config: Config;
 }) {
-  const { integrations } = options;
+  const { integrations, config } = options;
 
   return createTemplateAction<{
     repoUrl: string;
     description: string;
+    defaultBranch?: string;
     repoVisibility: 'private' | 'public';
     sourcePath?: string;
     enableLFS: boolean;
@@ -214,6 +217,11 @@ export function createPublishBitbucketAction(options: {
             title: 'Repository Visibility',
             type: 'string',
             enum: ['private', 'public'],
+          },
+          defaultBranch: {
+            title: 'Default Branch',
+            type: 'string',
+            description: `Sets the default branch on the repository. The default value is 'master'`,
           },
           sourcePath: {
             title:
@@ -245,6 +253,7 @@ export function createPublishBitbucketAction(options: {
       const {
         repoUrl,
         description,
+        defaultBranch = 'master',
         repoVisibility = 'private',
         enableLFS = false,
       } = ctx.input;
@@ -277,6 +286,11 @@ export function createPublishBitbucketAction(options: {
         apiBaseUrl,
       });
 
+      const gitAuthorInfo = {
+        name: config.getOptionalString('scaffolder.defaultAuthor.name'),
+        email: config.getOptionalString('scaffolder.defaultAuthor.email'),
+      };
+
       await initRepoAndPush({
         dir: getRepoSourceDirectory(ctx.workspacePath, ctx.input.sourcePath),
         remoteUrl,
@@ -288,7 +302,9 @@ export function createPublishBitbucketAction(options: {
             ? integrationConfig.config.appPassword
             : integrationConfig.config.token ?? '',
         },
+        defaultBranch,
         logger: ctx.logger,
+        gitAuthorInfo,
       });
 
       if (enableLFS && host !== 'bitbucket.org') {