+++ plugins/cost-insights/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @backstage/plugin-cost-insights
 
+## 0.11.1
+
+### Patch Changes
+
+- 0b4d00687: Replaced moment and dayjs with Luxon
+- Updated dependencies
+  - @backstage/core-components@0.1.4
+
 ## 0.11.0
 
 ### Minor Changes